moves collectible utils to handlers utils
