+++ internal/api/handlers/utils_test.go
@@ -1,4 +1,4 @@
-package utils_test
+package handlers
 
 import (
 	"context"
@@ -17,7 +17,7 @@ func TestFetchCollection_Success(t *testing.T) {
 	mockRPC := &utils.MockRPCService{}
 
 	account := &txnbuild.SimpleAccount{AccountID: "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF"}
-	collection, err := utils.FetchCollection(mockRPC, context.Background(), account, "CBIELTK6YBZJU5UP2WWQEUCYKLPU6AUNZ2BQ4WWFEIE3USCIHMXQDAMA")
+	collection, err := FetchCollection(mockRPC, context.Background(), account, "CBIELTK6YBZJU5UP2WWQEUCYKLPU6AUNZ2BQ4WWFEIE3USCIHMXQDAMA")
 
 	assert.NoError(t, err)
 	assert.NotNil(t, collection)
@@ -29,7 +29,7 @@ func TestFetchCollection_InvalidContractID(t *testing.T) {
 	mockRPC := &utils.MockRPCService{}
 	account := &txnbuild.SimpleAccount{AccountID: "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF"}
 
-	_, err := utils.FetchCollection(mockRPC, context.Background(), account, "INVALID")
+	_, err := FetchCollection(mockRPC, context.Background(), account, "INVALID")
 	assert.Error(t, err)
 }
 
@@ -39,7 +39,7 @@ func TestFetchCollection_SimulateInvocationError(t *testing.T) {
 	}
 
 	account := &txnbuild.SimpleAccount{AccountID: "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF"}
-	_, err := utils.FetchCollection(mockRPC, context.Background(), account, "CBIELTK6YBZJU5UP2WWQEUCYKLPU6AUNZ2BQ4WWFEIE3USCIHMXQDAMA")
+	_, err := FetchCollection(mockRPC, context.Background(), account, "CBIELTK6YBZJU5UP2WWQEUCYKLPU6AUNZ2BQ4WWFEIE3USCIHMXQDAMA")
 	assert.Error(t, err)
 }
 
@@ -55,7 +55,7 @@ func TestFetchCollectible_Success(t *testing.T) {
 	}
 
 	account := &txnbuild.SimpleAccount{AccountID: "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF"}
-	collectible, err := utils.FetchCollectible(mockRPC, context.Background(), account, "CBIELTK6YBZJU5UP2WWQEUCYKLPU6AUNZ2BQ4WWFEIE3USCIHMXQDAMA", tokenId)
+	collectible, err := FetchCollectible(mockRPC, context.Background(), account, "CBIELTK6YBZJU5UP2WWQEUCYKLPU6AUNZ2BQ4WWFEIE3USCIHMXQDAMA", tokenId)
 
 	assert.NoError(t, err)
 	assert.NotNil(t, collectible)
@@ -69,7 +69,7 @@ func TestFetchCollectible_InvalidTokenID(t *testing.T) {
 	mockRPC := &utils.MockRPCService{}
 	account := &txnbuild.SimpleAccount{AccountID: "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF"}
 
-	_, err := utils.FetchCollectible(mockRPC, context.Background(), account, "CBIELTK6YBZJU5UP2WWQEUCYKLPU6AUNZ2BQ4WWFEIE3USCIHMXQDAMA", "not-a-number")
+	_, err := FetchCollectible(mockRPC, context.Background(), account, "CBIELTK6YBZJU5UP2WWQEUCYKLPU6AUNZ2BQ4WWFEIE3USCIHMXQDAMA", "not-a-number")
 	assert.Error(t, err)
 }
 
@@ -79,6 +79,6 @@ func TestFetchCollectible_SimulateInvocationError(t *testing.T) {
 	}
 
 	account := &txnbuild.SimpleAccount{AccountID: "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF"}
-	_, err := utils.FetchCollectible(mockRPC, context.Background(), account, "CBIELTK6YBZJU5UP2WWQEUCYKLPU6AUNZ2BQ4WWFEIE3USCIHMXQDAMA", "1")
+	_, err := FetchCollectible(mockRPC, context.Background(), account, "CBIELTK6YBZJU5UP2WWQEUCYKLPU6AUNZ2BQ4WWFEIE3USCIHMXQDAMA", "1")
 	assert.Error(t, err)
 }
